crawler/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.10.0
anthropic>=0.20.0
jinja2>=3.1.0
schedule>=1.2.0
Pillow>=10.0.0
